compiles

+++ runtimes/manta/primitives/src/constants.rs
@@ -33,9 +33,9 @@ pub mod time {
 	// NOTE: Currently it is not possible to change the epoch duration after the chain has started.
 	//       Attempting to do so will brick block production.
 	pub const EPOCH_DURATION_IN_BLOCKS: BlockNumber = 10 * MINUTES;
-	pub const EPOCH_DURATION_IN_SLOTS: u64 = {
-		const SLOT_FILL_RATE: f64 = MILLISECS_PER_BLOCK as f64 / SLOT_DURATION as f64;
-		(EPOCH_DURATION_IN_BLOCKS as f64 * SLOT_FILL_RATE) as u64
+	pub const EPOCH_DURATION_IN_SLOTS: u32 = {
+		const SLOT_FILL_RATE: f32 = MILLISECS_PER_BLOCK as f32 / SLOT_DURATION as f32;
+		(EPOCH_DURATION_IN_BLOCKS as f32 * SLOT_FILL_RATE) as u32
 	};
 
 	// Time is measured by number of blocks.
